pyproject.toml
[tool.poetry]
name = "robot-code"
version = "0.1.0"
description = ""
authors = ["Alex Kempen <alex.bookreader@gmail.com>"]
readme = "README.md"
packages = [
    { include = "onshape_api" },
    { include = "backend" },
    { include = "featurescript" },
    { include = "robot_code" },
    { include = "backend_tools" },
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
json5 = "^0.9.14"
# Server
flask = {extras = ["async"], version = "^3.0.0"}
requests-oauthlib = "^1.3.1"
google-cloud-firestore = "^2.13.1"
gunicorn = "^21.2.0"
# Parsing
lark = {extras = ["interegular"], version="^1.1.9"}
semver = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
